# Conflicts:
#	app/src/main/java/com/gdg/zealicon2k25/data/remote/AuthApi.kt
#	app/src/main/java/com/gdg/zealicon2k25/presentation/ui/RegisterScreen.kt

+++ app/src/main/java/com/gdg/zealicon2k25/data/models/VerifyOtpReq.kt
@@ -2,5 +2,5 @@ package com.gdg.zealicon2k25.data.models
 
 data class VerifyOtpReq(
     val email:String ,
-    val otp:Int
+    val otp: Long
 )
